+++ oneEngine/oneGame/source/core-ext/animation/set/AnimationSet.h
@@ -26,7 +26,7 @@ public:
 		//animMap.push_back( newCurve );
 	}
 	void Add ( AnimationCurve<Matrix2x2>* );
-	void Add ( AnimationCurve<Vector3d>* );
+	void Add ( AnimationCurve<Vector3f>* );
 	void Add ( AnimationCurve<Vector2f>* );
 	void Add ( AnimationCurve<Real>* );
 	void Add ( AnimationCurve<XTransform>* newCurve )
