+++ oneEngine/oneGame/source/core/math/Line.h
@@ -4,23 +4,23 @@
 #define _STRUCT_LINE_H_
 
 // Includes
-#include "Vector3d.h"
+#include "Vector3.h"
 #include "Ray.h"
 
 // Struct definition
 struct Line
 {
-	Vector3d start;
-	Vector3d end;
+	Vector3f start;
+	Vector3f end;
 public:
 	Line ( void ) {
-		start	= Vector3d( 0,0,0 );
-		end		= Vector3d( 0,0,1 );
+		start	= Vector3f( 0,0,0 );
+		end		= Vector3f( 0,0,1 );
 	}
 	Line ( const Line& l ) {
 		*this = l;
 	}
-	Line ( const Vector3d& startpoint, const Vector3d& endpoint )
+	Line ( const Vector3f& startpoint, const Vector3f& endpoint )
 	{
 		start	= startpoint;
 		end		= endpoint;
@@ -30,7 +30,7 @@ public:
 		start	= ray.pos;
 		end		= ray.pos + ray.dir;
 	}
-	Vector3d midpoint ( void ) const
+	Vector3f midpoint ( void ) const
 	{
 		return ( (start+end)*0.5f );
 	}
