+++ oneEngine/oneGame/source/core/math/Plane.h
@@ -3,7 +3,7 @@
 
 // Includes
 #include "core/types/float.h"
-#include "Vector3d.h"
+#include "Vector3.h"
 
 // Class Def
 namespace core
@@ -13,15 +13,15 @@ namespace core
 		class Plane
 		{
 		public:
-			Vector3d n;
+			Vector3f n;
 			Real d;
 
 		public:
 			// Constructors
 			Plane ( void );
-			Plane ( Vector3d * );
-			void ConstructFromPoints ( Vector3d * );
-			void ConstructFromPoints ( Vector3d const&, Vector3d const&, Vector3d const& );
+			Plane ( Vector3f * );
+			void ConstructFromPoints ( Vector3f * );
+			void ConstructFromPoints ( Vector3f const&, Vector3f const&, Vector3f const& );
 		};
 	}
 }
